email auth working
