+++ src/Components/modal/ModalScreen.css
@@ -1,0 +1,34 @@
+.otpIcon {
+  height: 140px;
+  display: block;
+  width: 208px;
+  background-position: -250px -365px;
+  background-repeat: no-repeat;
+  margin: 0 15%;
+}
+.otpInput {
+  border-width: 1px;
+  border: 1px solid #d2d2d2;
+  font-size: 12px;
+  line-height: 20px;
+  outline: none;
+  text-align: center;
+  width: 50%;
+  background-color: #fbfbfb !important;
+  padding: 2% 0%;
+}
+.otpDiv {
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+  justify-content: center;
+  align-items: center;
+}
+.resendOtp{
+    color: #25a8ed;
+    font-size: 12px;
+    cursor: pointer;
+    background-color: white;
+    border: none;
+    margin:4% 0% 0% 0%
+}
